Adding _process_functions.py
