+++ pynapple/core/_jitted_functions.py
@@ -497,106 +497,6 @@ def jitcontinuous_perievent(
     return new_data_array
 
 
-@jit(nopython=True)
-def jitperievent_trigger_average(
-    time_array,
-    count_array,
-    time_target_array,
-    data_target_array,
-    starts,
-    ends,
-    windows,
-    binsize,
-):
-    T = time_array.shape[0]
-    N = count_array.shape[1]
-    N_epochs = len(starts)
-
-    idx, count = jitrestrict_with_count(time_target_array, starts, ends)
-    time_target_array = time_target_array[idx]
-    data_target_array = data_target_array[idx]
-
-    max_count = np.cumsum(count)
-
-    new_data_array = np.full(
-        (int(windows.sum()) + 1, count_array.shape[1], *data_target_array.shape[1:]),
-        0.0,
-    )
-
-    t = 0  # count events
-
-    hankel_array = np.zeros((new_data_array.shape[0], *data_target_array.shape[1:]))
-
-    for k in range(N_epochs):
-        if count[k] > 0:
-            t_start = t
-            maxi = max_count[k]
-            i = maxi - count[k]
-
-            while t < T:
-                lbound = time_array[t]
-                rbound = np.round(lbound + binsize, 9)
-
-                if time_target_array[i] < rbound:
-                    i_start = i
-                    i_stop = i
-
-                    while i_stop < maxi:
-                        if time_target_array[i_stop] < rbound:
-                            i_stop += 1
-                        else:
-                            break
-
-                    while i_start < i_stop - 1:
-                        if time_target_array[i_start] < lbound:
-                            i_start += 1
-                        else:
-                            break
-                    v = np.sum(data_target_array[i_start:i_stop], 0) / float(
-                        i_stop - i_start
-                    )
-
-                    checknan = np.sum(v)
-                    if not np.isnan(checknan):
-                        hankel_array[-1] = v
-
-                if t - t_start >= windows[1]:
-                    for n in range(N):
-                        new_data_array[:, n] += (
-                            hankel_array * count_array[t - windows[1], n]
-                        )
-
-                # hankel_array = np.roll(hankel_array, -1, axis=0)
-                hankel_array[0:-1] = hankel_array[1:]
-                hankel_array[-1] = 0.0
-
-                t += 1
-
-                i = i_start
-
-                if t == T or time_array[t] > ends[k]:
-                    if t - t_start > windows[1]:
-                        for j in range(windows[1]):
-                            for n in range(N):
-                                new_data_array[:, n] += (
-                                    hankel_array * count_array[t - windows[1] + j, n]
-                                )
-
-                            # hankel_array = np.roll(hankel_array, -1, axis=0)
-                            hankel_array[0:-1] = hankel_array[1:]
-                            hankel_array[-1] = 0.0
-
-                    hankel_array *= 0.0
-                    break
-
-    total = np.sum(count_array, 0)
-    for n in range(N):
-        if total[n] > 0.0:
-            new_data_array[:, n] /= total[n]
-
-    return new_data_array
-
-
 ################################
 # IntervalSet functions
 ################################
